finalise jokes and fibonacci with a stateful async iterable

+++ src/schema.ts
@@ -11,7 +11,7 @@ export const typeDefs = gql`
 
     type Subscription {
         jokes: String!
-        fibonacci: Int!
+        fibonacci: Float!
         listen: String!
     }
 `;
